Backend: Move WSLKubernetesBackend to new file.
This is the WSL equvialent of 3f23427e7ec9acabc1008bf547d053ccd3bc5ac4.
